chore(api): fix golangci-lint — exitAfterDefer + gosec int→int32
Three issues surfaced by the Go CI lint stage:

- cmd/api/main.go boot path used log.Fatal (and a bare-error exit)
  while having defers for pool/cache/sentry Close/Flush. gocritic's
  exitAfterDefer is right — those defers never run past a fatal exit.
  Restructured so every init error logs + os.Exit(1) BEFORE the
  corresponding defer is registered, and the boot-time config failure
  path switched from boot.Fatal to Error + os.Exit(1) for consistency.
- Renamed the local `a` variable that shadowed the `app` package
  import.
- handlers/internal_ai_usage.go aiUsageRequest.InputTokens /
  OutputTokens typed as int were cast to int32 for the sqlc params
  → gosec G115 overflow risk. Changed the struct fields to int32
  directly; JSON doesn't care, and a single model call is never
  anywhere near 2B tokens.

No behaviour change.

+++ apps/api/internal/handlers/internal_ai_usage.go
@@ -48,8 +48,8 @@ type aiUsageRequest struct {
 	UserID         uuid.UUID       `json:"user_id"`
 	ConversationID *uuid.UUID      `json:"conversation_id"`
 	Model          string          `json:"model"`
-	InputTokens    int             `json:"input_tokens"`
-	OutputTokens   int             `json:"output_tokens"`
+	InputTokens    int32           `json:"input_tokens"`
+	OutputTokens   int32           `json:"output_tokens"`
 	CostUSD        decimal.Decimal `json:"cost_usd"`
 }
 
@@ -143,8 +143,8 @@ func writeAIUsageTx(ctx context.Context, deps *app.Deps, req aiUsageRequest) err
 		UserID:         req.UserID,
 		ConversationID: req.ConversationID,
 		Model:          req.Model,
-		InputTokens:    int32(req.InputTokens),
-		OutputTokens:   int32(req.OutputTokens),
+		InputTokens:    req.InputTokens,
+		OutputTokens:   req.OutputTokens,
 		CostUsd:        req.CostUSD,
 	}); err != nil {
 		return fmt.Errorf("insert ai_usage: %w", err)
